requirements.txt
astropy==6.1.4
astropy-iers-data==0.2024.10.28.0.34.7
astroquery==0.4.7
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
keyring==25.5.0
more-itertools==10.5.0
numpy==2.1.2
packaging==24.1
psycopg2==2.9.10
pycparser==2.22
pyerfa==2.0.1.4
python-dotenv==1.0.1
pyvo==1.5.3
PyYAML==6.0.2
requests==2.32.3
SecretStorage==3.3.3
six==1.16.0
soupsieve==2.6
urllib3==2.2.3
webencodings==0.5.1
zipp==3.20.2
